Use a util func to get current config value
